requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
gitdb==4.0.5
GitPython==3.1.3
idna==2.9
PyYAML==5.3.1
requests==2.24.0
smmap==3.0.4
urllib3==1.25.9
